requirements.txt
Flask
python-dotenv
flask-cors
langchain
langchain-community
langchain-openai
openai
pinecone-client
langchain-pinecone 
pypdf
cors